% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayr_payload.R
\name{example_displayr_payload}
\alias{example_displayr_payload}
\title{Build Example Displayr Payload}
\usage{
example_displayr_payload()
}
\value{
Named list payload from \code{\link[=build_displayr_payload]{build_displayr_payload()}}.
}
\description{
Convenience helper for local/dev testing.
}
