% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamental_examples.R
\name{example_fundamental_inputs}
\alias{example_fundamental_inputs}
\title{Example Inputs for Local Testing}
\usage{
example_fundamental_inputs()
}
\value{
Named list containing \code{index_data} and \code{user_data}.
}
\description{
Returns example index and user datasets.
}
