% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamental_examples.R
\name{example_ohep_snapshot}
\alias{example_ohep_snapshot}
\title{Example Prepared Snapshot for Local Testing}
\usage{
example_ohep_snapshot()
}
\value{
Named list of prepared snapshot marts.
}
\description{
Runs \code{\link[=prep_ohep_snapshot]{prep_ohep_snapshot()}} against package extdata and returns marts.
}
