% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamental_helpers.R
\name{get_delta_pill}
\alias{get_delta_pill}
\title{Get Delta Pill Class}
\usage{
get_delta_pill(delta_value)
}
\arguments{
\item{delta_value}{Numeric change value.}
}
\value{
One of \code{dp-pos}, \code{dp-neg}, \code{dp-neu}.
}
\description{
Convenience helper that maps a delta value to pill classes.
}
