% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamental_helpers.R
\name{get_sentiment_class}
\alias{get_sentiment_class}
\title{Get Sentiment Class}
\usage{
get_sentiment_class(score)
}
\arguments{
\item{score}{Numeric percentage score.}
}
\value{
One of \code{bg-agree}, \code{bg-neutral}, \code{bg-disagree}.
}
\description{
Convenience helper that maps percentage values to sentiment classes used by
the renderer.
}
