% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohep_snapshot.R
\name{load_ohep_snapshot}
\alias{load_ohep_snapshot}
\title{Load OHEP Snapshot}
\usage{
load_ohep_snapshot(path_or_tables)
}
\arguments{
\item{path_or_tables}{Either a directory path containing snapshot files or
an already-loaded named list of snapshot tables.}
}
\value{
Named list of snapshot tables.
}
\description{
Load OHEP Snapshot
}
